feat(indexer): channel info, zeroize felt, privacy pool types
