+++ crates/discovery-core/src/lib.rs
@@ -1,3 +1,4 @@
 pub mod mock_backend;
 pub mod storage;
 pub mod storage_slots;
+pub mod types;
