+++ crates/discovery-core/src/storage.rs
@@ -7,35 +7,7 @@ use starknet_types_core::felt::Felt;
 use thiserror::Error;
 
 use crate::storage_slots;
-
-/// Ciphertext for an ECDH-based encryption of channel data.
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub struct EncChannelInfo {
-    /// Ephemeral ECDH public key x-coordinate (rG.x).
-    pub ephemeral_pubkey: Felt,
-    /// Encrypted channel key.
-    pub enc_channel_key: Felt,
-    /// Encrypted sender address.
-    pub enc_sender_addr: Felt,
-}
-
-/// Ciphertext for an ECDH-based encryption of private key.
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub struct EncPrivateKey {
-    /// Ephemeral ECDH public key x-coordinate (rG.x).
-    pub ephemeral_pubkey: Felt,
-    /// Encrypted private key.
-    pub enc_private_key: Felt,
-}
-
-/// Encrypted subchannel info.
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub struct EncSubchannelInfo {
-    /// Salt generated by the sender for one-time key usage.
-    pub salt: Felt,
-    /// Encrypted token.
-    pub enc_token: Felt,
-}
+use crate::types::{EncChannelInfo, EncPrivateKey, EncSubchannelInfo};
 
 /// Errors that can occur during storage operations.
 #[derive(Debug, Error)]
@@ -72,7 +44,7 @@ pub trait StorageSnapshot: IViews {
     fn block_id(&self) -> BlockId;
 }
 
-/// Privacy contract view methods
+/// Privacy contract view methods.
 #[async_trait]
 pub trait IViews: Send + Sync {
     /// Checks if a channel with the given ID exists.
@@ -123,7 +95,7 @@ pub trait RawStorageAccess: Send + Sync {
     async fn read_slots(&self, slots: Vec<Felt>) -> Result<Vec<Felt>, StorageError>;
 }
 
-/// Blanket implementation of `PrivacyContractStorage` for any type implementing `RawStorageAccess`.
+/// Blanket implementation of `IViews` for any type implementing `RawStorageAccess`.
 #[async_trait]
 impl<T: RawStorageAccess> IViews for T {
     async fn channel_exists(&self, channel_id: Felt) -> Result<bool, StorageError> {
