+++ crates/discovery-core/src/types.rs
@@ -1,0 +1,74 @@
+//! Privacy pool specific types.
+//!
+//! This module contains the data structures used by the privacy pool contract,
+//! including encrypted ciphertext types and decrypted plaintext types.
+
+use starknet_types_core::felt::Felt;
+
+/// Trait for securely zeroing sensitive data from memory.
+///
+/// This is a simplified version of the `zeroize` crate's `Zeroize` trait,
+/// implemented specifically for `Felt` to avoid the orphan rule.
+pub trait Zeroize {
+    /// Securely zeros the memory containing this value.
+    fn zeroize(&mut self);
+}
+
+impl Zeroize for Felt {
+    fn zeroize(&mut self) {
+        // Overwrite with zero
+        *self = Felt::ZERO;
+        // Use a compiler fence to prevent optimization from removing the write
+        std::sync::atomic::compiler_fence(std::sync::atomic::Ordering::SeqCst);
+    }
+}
+
+/// Ciphertext for an ECDH-based encryption of channel data.
+#[derive(Debug, Clone, Copy, PartialEq, Eq)]
+pub struct EncChannelInfo {
+    /// Ephemeral ECDH public key x-coordinate (rG.x).
+    pub ephemeral_pubkey: Felt,
+    /// Encrypted channel key.
+    pub enc_channel_key: Felt,
+    /// Encrypted sender address.
+    pub enc_sender_addr: Felt,
+}
+
+/// Ciphertext for an ECDH-based encryption of private key.
+#[derive(Debug, Clone, Copy, PartialEq, Eq)]
+pub struct EncPrivateKey {
+    /// Ephemeral ECDH public key x-coordinate (rG.x).
+    pub ephemeral_pubkey: Felt,
+    /// Encrypted private key.
+    pub enc_private_key: Felt,
+}
+
+/// Encrypted subchannel info.
+#[derive(Debug, Clone, Copy, PartialEq, Eq)]
+pub struct EncSubchannelInfo {
+    /// Salt generated by the sender for one-time key usage.
+    pub salt: Felt,
+    /// Encrypted token.
+    pub enc_token: Felt,
+}
+
+/// Decrypted channel information.
+#[derive(Debug, Clone, Copy, PartialEq, Eq)]
+pub struct ChannelInfo {
+    /// The channel key.
+    pub channel_key: Felt,
+    /// The sender's address.
+    pub sender_addr: Felt,
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    #[test]
+    fn test_zeroize_felt() {
+        let mut secret = Felt::from(42u64);
+        secret.zeroize();
+        assert_eq!(secret, Felt::ZERO);
+    }
+}
